fix: 업그레이드 적용안되는 점 수정

+++ Assets/01.Scripts/Tree/TreeOutgame/Feature/01.Domain/Upgrade/IReadonlyUpgrade.cs
@@ -5,7 +5,7 @@
     string Description { get; }
     int Level { get; }
     Currency Cost { get; }
-    double Damage { get; }
+    int Damage { get; }
     bool IsMaxLevel { get; }
     UnityEngine.Sprite Icon { get; }
 }
